# mehrdadhub-openapi-v0.json
{
	"openapi": "3.0.0",
	"paths": {
		"/connection-test": {
			"get": {
				"operationId": "connectionTest",
				"parameters": [],
				"responses": {
					"200": {
						"description": ""
					}
				},
				"tags": [
					"Root"
				]
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "tempLogin",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				},
				"tags": [
					"Auth"
				]
			}
		},
		"/users/register": {
			"post": {
				"operationId": "register",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": ""
					}
				},
				"tags": [
					"User"
				]
			}
		},
		"/users/me": {
			"get": {
				"operationId": "getUser",
				"parameters": [],
				"responses": {
					"200": {
						"description": ""
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/task": {
			"post": {
				"operationId": "create",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTaskDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": ""
					}
				}
			},
			"get": {
				"operationId": "findAll",
				"parameters": [],
				"responses": {
					"200": {
						"description": ""
					}
				}
			}
		},
		"/task/{id}": {
			"get": {
				"operationId": "findOne",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"patch": {
				"operationId": "update",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTaskDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"delete": {
				"operationId": "remove",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": ""
					}
				}
			}
		},
		"/task/{id}/image": {
			"patch": {
				"operationId": "changeImage",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangeImageDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			}
		}
	},
	"info": {
		"title": "MehrdadHub API",
		"description": "This is documentation for MehrdadHub rest API service",
		"version": "0.1.0",
		"contact": {}
	},
	"tags": [],
	"servers": [
		{
			"url": "http://localhost:5050/v0"
		}
	],
	"components": {
		"securitySchemes": {
			"bearer": {
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"type": "http"
			}
		},
		"schemas": {
			"LoginDto": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string",
						"description": "The email address of the user",
						"example": "user@example.com"
					},
					"password": {
						"type": "string",
						"description": "The password for the user account"
					}
				},
				"required": [
					"email",
					"password"
				]
			},
			"RegisterDto": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string",
						"description": "The email address of the user registering",
						"example": "user@example.com"
					},
					"password": {
						"type": "string",
						"description": "The password for the user account",
						"example": "yourStrong(!)Password",
						"minLength": 8
					},
					"displayName": {
						"type": "string",
						"description": "Display name of the user",
						"example": "John Doe"
					}
				},
				"required": [
					"email",
					"password",
					"displayName"
				]
			},
			"CreateTaskDto": {
				"type": "object",
				"properties": {
					"title": {
						"type": "string",
						"example": "My special task",
						"description": "A proper name for the task",
						"default": "New Task"
					},
					"description": {
						"type": "string",
						"example": "This is some description about my task",
						"description": "A proper description for the task"
					},
					"color": {
						"type": "string",
						"example": "blue",
						"description": "Color of the task card in UI (any proper CSS color format)"
					}
				},
				"required": [
					"title"
				]
			},
			"UpdateTaskDto": {
				"type": "object",
				"properties": {}
			},
			"ChangeImageDto": {
				"type": "object",
				"properties": {
					"imgUrl": {
						"type": "string",
						"description": "Url of the uploaded image"
					}
				},
				"required": [
					"imgUrl"
				]
			}
		}
	}
}